Fix vyper contracts crashing flattener
