+++ packages/discovery/src/index.ts
@@ -52,7 +52,6 @@ export { NoCache } from './discovery/provider/NoCache'
 export {
   buildSimilarityHashmap,
   estimateSimilarity,
-  removeComments,
 } from './flatten/utils'
 export { EtherscanClient } from './utils/EtherscanClient'
 export { getErrorMessage } from './utils/getErrorMessage'
